requirements.txt
wheel==0.38.4
aiofiles==22.1.0
fastapi==0.88.0
face_recognition==1.3.0
gunicorn==20.1.0
numpy==1.23.5
opencv-python==4.6.0.66
scipy==1.9.3
requests==2.28.1
starlette==0.22.0
starlette_context==0.3.5
passlib[bcrypt]==1.7.4
pydantic==1.10.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
uvicorn[standard]==0.20.0
pypandoc==1.10
pyyaml==6.0
